Works on stable using proc macro hack
