+++ Cargo.toml
@@ -1,33 +1,6 @@
-[package]
-name = "wchar"
-version = "0.2.0"
-authors = ["Juici <juicy66173@gmail.com>"]
-description = "Procedural macros for compile time UTF-16 strings"
+[workspace]
 
-documentation = "https://docs.rs/wchar"
-homepage = "https://github.com/Juici/wchar-rs"
-repository = "https://github.com/Juici/wchar-rs"
-readme = "README.md"
-license = "MIT"
-
-keywords = ["wide", "string", "wchar", "utf16"]
-categories = ["text-processing"]
-
-include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
-
-[badges]
-appveyor = { repository = "Juici/wchar-rs" }
-travis-ci = { repository = "Juici/wchar-rs" }
-maintenance = { status = "passively-maintained" }
-
-[lib]
-proc-macro = true
-
-[dependencies.syn]
-version = "0.14"
-default-features = false
-features = ["derive", "parsing"]
-
-[dependencies.proc-macro2]
-version = "0.4"
-features = ["nightly"]
+members = [
+    "wchar-rs",
+    "wchar-rs-macro",
+]
